requirements.txt
torch>=2.0.0
transformers>=4.44
accelerate>=0.20
datasets>=2.10
scikit-learn>=1.0
safetensors>=0.4.0
peft>=0.4.0
trl>=0.7.0
pandas>=1.3
pillow>=9.0
matplotlib>=3.5
numpy>=1.20
